+++ Source/Magick.NET.Web/Properties/AssemblyInfo.cs
@@ -55,7 +55,11 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyVersion("7.0.0.0")]
 [assembly: AssemblyFileVersion("7.0.5.502")]
 [assembly: ComVisible(false)]
+#if Q16
+[assembly: CLSCompliant(false)]
+#else
 [assembly: CLSCompliant(true)]
+#endif
 
 [assembly: InternalsVisibleTo("Magick.NET.Tests, PublicKey=" +
 "002400000480000094000000060200000024000052534131000400000100010041848921d7f5c3" +